pyproject.toml
[project]
name = "totem-server"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = "==3.13.*"
dependencies = [
  "argon2-cffi<26",
  "cssselect<2",
  "environs[django]<16",
  "google-api-python-client<3",
  "google-auth-oauthlib<2",
  "lxml<7",
  "markdown<3.11.0",
  "mrml<0.3.0",
  "Pillow<13",
  "posthog<8.0.0",
  "psycopg[binary]<4",
  "python-dateutil<3",
  "pytz",
  "qrcode<9",
  "requests<3",
  "sentry-sdk[django]<3",
  "whitenoise<7",
  # Django dependencies
  "django-anymail<15",
  "django-auditlog<4",
  "django-cors-headers<5",
  "django-imagekit<7.0.0",
  "django-impersonate<2",
  "django-ninja<2",
  "django-storages[s3]<2",
  "django-taggit<7",
  "django-timezone-field<8",
  "tzdata",
  "django<6.1",
  "firebase-admin>=6.8.0",
  "tenacity>=9.1.2",
  "livekit-api<2",
  "granian<3",
  "dynimg<=0.2.0",
  "django-oauth-toolkit>=3.2.0",
]

[dependency-groups]
dev = [
  "ipdb==0.13.13",
  "watchfiles<2",
  "uv",

  # Testing
  "mypy<2",
  "django-stubs[compatible-mypy]<7",
  "pytest<10",
  "pytest-sugar<2",
  "pytest-socket<1",

  # Code quality
  "coverage<8.0.0",
  "pip-tools<8",
  "pre-commit<5",
  "ruff",

  # Django
  "django-debug-toolbar<7",
  "django-extensions<5",
  "factory-boy<4",
  "pytest-cov<8",
  "pytest-django<5",
  "pytest-xdist<4",
  "debugpy",
]


# ==== pytest ====
[tool.pytest.ini_options]
minversion = "6.0"
addopts = "--ds=config.settings.test --reuse-db"
python_files = ["tests.py", "test_*.py"]
filterwarnings = [
  # "ignore::pydantic.warnings.PydanticDeprecatedSince20",
  "ignore:::httplib2.*",
]

# ==== Coverage ====
[tool.coverage.run]
include = ["totem/**"]
omit = ["*/migrations/*", "*/tests/*"]
disable_warnings = ["include-ignored"]


# ==== mypy ====
[tool.mypy]
python_version = "3.13"
check_untyped_defs = true
ignore_missing_imports = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_unused_configs = true
plugins = ["mypy_django_plugin.main"]

[[tool.mypy.overrides]]
# Django migrations should not produce any errors:
module = "*.migrations.*"
ignore_errors = true

[tool.django-stubs]
django_settings_module = "config.settings.test"


# ==== Ruff =====
[tool.ruff]
# Exclude a variety of commonly ignored directories.
exclude = ["migrations", ".venv"]
line-length = 120

[tool.ruff.lint]
# Enable pycodestyle (`E`) and Pyflakes (`F`) codes by default.
ignore = ["E501", "DJ001", "DJ008", "DJ012"]
select = [
  # pycodestyle
  "E",
  # Pyflakes
  "F",
  # # pyupgrade
  # "UP",
  # # flake8-bugbear
  # "B",
  # isort
  "I",
  # django
  "DJ",
]

[tool.ruff.lint.mccabe]
# Unlike Flake8, default to a complexity level of 10.
max-complexity = 10

# ==== isort ====
[tool.ruff.lint.isort]
known-first-party = ["totem", "config"]
